tgbot/requirements.txt
psycopg2-binary
python-dotenv
numpy
catboost
python-telegram-bot